+++ src/OpenConext/EngineBlock/Metadata/Entity/ServiceProvider.php
@@ -103,6 +103,13 @@ class ServiceProvider extends AbstractRole
     /**
      * @var bool
      *
+     * @ORM\Column(name="attribute_aggregation_required", type="boolean")
+     */
+    public $attributeAggregationRequired;
+
+    /**
+     * @var bool
+     *
      * @ORM\Column(name="requesterid_required", type="boolean")
      */
     public $requesteridRequired;
@@ -165,6 +172,7 @@ class ServiceProvider extends AbstractRole
      * @param null $requestedAttributes
      * @param bool $skipDenormalization
      * @param bool $policyEnforcementDecisionRequired
+     * @param bool $attributeAggregationRequired
      * @param bool $requesteridRequired
      * @param string $manipulation
      * @param AttributeReleasePolicy $attributeReleasePolicy
@@ -211,6 +219,7 @@ class ServiceProvider extends AbstractRole
         $requestedAttributes = null,
         $skipDenormalization = false,
         $policyEnforcementDecisionRequired = false,
+        $attributeAggregationRequired = false,
         $requesteridRequired = false,
         $manipulation = '',
         AttributeReleasePolicy $attributeReleasePolicy = null,
@@ -258,6 +267,7 @@ class ServiceProvider extends AbstractRole
         $this->requestedAttributes = $requestedAttributes;
         $this->skipDenormalization = $skipDenormalization;
         $this->policyEnforcementDecisionRequired = $policyEnforcementDecisionRequired;
+        $this->attributeAggregationRequired = $attributeAggregationRequired;
         $this->requesteridRequired = $requesteridRequired;
         $this->supportUrlEn = $supportUrlEn;
         $this->supportUrlNl = $supportUrlNl;